Kullanıcılara verilen cevaplar veritabanından çekildi. Eğer öyle bir kullanıcı yoksa chatGPt ye sorulacak şekilde ayarlanacak

+++ AIService/CQRS/Register/Dto/UserDto.cs
@@ -6,8 +6,8 @@
         public string Surname { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
-        public int Kilo { get; set; }
-        public int Height { get; set; }
+        public double Kilo { get; set; }
+        public double Height { get; set; }
         public int Age { get; set; }
         public bool Gender { get; set; }
     }
